change from float to int

+++ kalkulator.py
@@ -10,18 +10,18 @@ def bangun_2d(): # function untuk mempermudah dalam memilih kalkulator 2d
     pilihan = input("Silahkan di pilih: ")
 
     if pilihan == "1": # Menghitung luas persegi
-        sisi = float(input("Masukkan panjang sisi: "))
+        sisi = int(input("Masukkan panjang sisi: "))
         print("Luas Persegi: ", sisi ** 2)
         
     elif pilihan == "2": # Menghitung luas segitiga
-        alas = float(input("Masukkan panjang alas: "))
-        tinggi = float(input("Masukkan tinggi: "))
+        alas = int(input("Masukkan panjang alas: "))
+        tinggi = int(input("Masukkan tinggi: "))
         print("Luas Segitiga: ",0.5 * alas * tinggi)
         
     elif pilihan == "3": # Menghitung luas trapesium
-        sisi_atas = float(input("Masukkan panjang sisi atas: "))
-        sisi_bawah = float(input("Masukkan panjang sisi bawah: "))
-        tinggi = float(input("Masukkan tinggi: "))
+        sisi_atas = int(input("Masukkan panjang sisi atas: "))
+        sisi_bawah = int(input("Masukkan panjang sisi bawah: "))
+        tinggi = int(input("Masukkan tinggi: "))
         print("Luas Trapesium: ", 0.5 * (sisi_atas + sisi_bawah) * tinggi)
         
     elif pilihan == "4": # Menghitung luas jari- jari linkaran
@@ -29,8 +29,8 @@ def bangun_2d(): # function untuk mempermudah dalam memilih kalkulator 2d
         print("Luas Lingkaran: ", pi * jari_jari ** 2)
         
     elif pilihan == "5": # Menghitung luas persegi panjang
-        panjang = float(input("Masukkan panjang: "))
-        lebar = float(input("Masukkan lebar: "))
+        panjang = int(input("Masukkan panjang: "))
+        lebar = int(input("Masukkan lebar: "))
         print("Luas Persegi Panjang: ", panjang * lebar)
         
     else: # tidak valid
@@ -46,27 +46,27 @@ def bangun_3d(): # function untuk mempermudah dalam memilih kalkulator 3d
     pilihan = input("Silahkan di Pilih: ")
 
     if pilihan == "1": # Menghitung volume bola
-        jari_jari = float(input("Masukkan jari-jari: "))
+        jari_jari = int(input("Masukkan jari-jari: "))
         print("Volume Bola: ", (4/3) * pi * jari_jari ** 3)
         
     elif pilihan == "2": # Menghitung volume kubus
-        sisi = float(input("Masukkan panjang sisi: "))
+        sisi = int(input("Masukkan panjang sisi: "))
         print("Volume Kubus: ", sisi ** 3)
         
     elif pilihan == "3": # Menghitung volume balok
-        panjang = float(input("Masukkan panjang: "))
-        lebar = float(input("Masukkan lebar: "))
-        tinggi = float(input("Masukkan tinggi: "))
+        panjang = int(input("Masukkan panjang: "))
+        lebar = int(input("Masukkan lebar: "))
+        tinggi = int(input("Masukkan tinggi: "))
         print("Volume Balok: ", panjang * lebar * tinggi)
         
     elif pilihan == "4": # Menghitung volume tabung
-        jari_jari = float(input("Masukkan jari-jari: "))
-        tinggi = float(input("Masukkan tinggi: "))
+        jari_jari = int(input("Masukkan jari-jari: "))
+        tinggi = int(input("Masukkan tinggi: "))
         print("Volume Tabung: ", pi * jari_jari ** 2 * tinggi)
         
     elif pilihan == "5": # Menghitung volume kerucut
-        jari_jari = float(input("Masukkan jari-jari: "))
-        tinggi = float(input("Masukkan tinggi: "))
+        jari_jari = int(input("Masukkan jari-jari: "))
+        tinggi = int(input("Masukkan tinggi: "))
         print("Volume Kerucut: ", (1/3) * pi * jari_jari ** 2 * tinggi)
         
     else: # tidak valid
